+++ src/reducers/blockchain.js
@@ -5,11 +5,15 @@ import {
   setGnosisInitialized,
   setGasCost,
   setGasPrice,
+  setEtherTokens,
+} from 'actions/blockchain'
+
+import {
   setActiveProvider,
   registerProvider,
   updateProvider,
-  setEtherTokens,
-} from 'actions/blockchain'
+} from 'actions/providers'
+
 import { GAS_COST } from 'utils/constants'
 
 const INITIAL_PROVIDER_STATE = {
